requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.5
chromadb==0.4.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0